Move strings to ember-strings to not have overlap in names
