+++ src/app/util/ember-strings.cpp
