+++ src/app/util/ember-strings.h
